Increased servos precision
Usually we write to the servo library an integer value, our angle in degrees. However doing it this way limits the potential of a servo motor. In fact, servos are controlled with a pulse in microseconds usually in the range 544 to 2400, which allows a finer control respect to the 0 to 180 range that the library offers us.

Moreover doing this way we can actually use the whole range supported by our motors, of around 203 degrees.